requirements-win.txt
flask
flask-cors
flask-jwt-extended
sqlalchemy 
psycopg2
cryptography
python-dotenv
Authlib
requests
